+++ src/FluentlyHttpClient/FluentHttpClientFactory.cs
@@ -15,6 +15,13 @@ namespace FluentlyHttpClient
 		/// <param name="identifier">identifier to set.</param>
 		/// <returns>Returns a new HTTP client builder.</returns>
 		FluentHttpClientBuilder CreateBuilder(string identifier);
+		
+		/// <summary>
+		/// Configure defaults for <see cref="FluentHttpClientBuilder"/> which every new one uses.
+		/// </summary>
+		/// <param name="configure">Configuration function.</param>
+		/// <returns>Returns client factory for chaining.</returns>
+		IFluentHttpClientFactory ConfigureDefaults(Action<FluentHttpClientBuilder> configure);
 
 		/// <summary>
 		/// Get <see cref="IFluentHttpClient"/> registered by identifier.
@@ -60,6 +67,7 @@ namespace FluentlyHttpClient
 	{
 		private readonly IServiceProvider _serviceProvider;
 		private readonly Dictionary<string, IFluentHttpClient> _clientsMap = new Dictionary<string, IFluentHttpClient>();
+		private Action<FluentHttpClientBuilder> _configure;
 
 		/// <summary>
 		/// Initializes a new instance of <see cref="FluentHttpClientFactory"/>.
@@ -78,8 +86,24 @@ namespace FluentlyHttpClient
 		public FluentHttpClientBuilder CreateBuilder(string identifier)
 		{
 			var clientBuilder = ActivatorUtilities.CreateInstance<FluentHttpClientBuilder>(_serviceProvider, this)
-				.Withdentifier(identifier);
+				.Withdentifier(identifier)
+				.WithUserAgent("fluently")
+				.WithTimeout(15);
+
+			_configure?.Invoke(clientBuilder);
+
 			return clientBuilder;
+		}
+
+		/// <summary>
+		/// Configure defaults for <see cref="FluentHttpClientBuilder"/> which every new one uses.
+		/// </summary>
+		/// <param name="configure">Configuration function.</param>
+		/// <returns>Returns client factory for chaining.</returns>
+		public IFluentHttpClientFactory ConfigureDefaults(Action<FluentHttpClientBuilder> configure)
+		{
+			_configure = configure;
+			return this;
 		}
 
 		/// <summary>
@@ -109,9 +133,7 @@ namespace FluentlyHttpClient
 
 			if (Has(options.Identifier))
 				throw new ClientBuilderValidationException($"FluentHttpClient '{options.Identifier}' is already registered.");
-
-			SetDefaultOptions(options);
-
+			
 			if (string.IsNullOrEmpty(options.BaseUrl))
 				throw ClientBuilderValidationException.FieldNotSpecified(nameof(options.BaseUrl));
 
@@ -154,16 +176,5 @@ namespace FluentlyHttpClient
 		/// <param name="identifier">Identifier to check.</param>
 		/// <returns>Returns true when already exists.</returns>
 		public bool Has(string identifier) => _clientsMap.ContainsKey(identifier);
-
-		/// <summary>
-		/// Merge default options with the specified options.
-		/// </summary>
-		/// <param name="options">Options to check and merge with.</param>
-		protected void SetDefaultOptions(FluentHttpClientOptions options)
-		{
-			if (options == null) throw new ArgumentNullException(nameof(options));
-			if (options.Timeout == TimeSpan.Zero)
-				options.Timeout = TimeSpan.FromSeconds(15);
-		}
 	}
 }