+++ src/FluentlyHttpClient/FluentHttpRequest.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Net.Http;
+using System.Net.Http.Formatting;
 using System.Net.Http.Headers;
 using System.Threading;
 using System.Threading.Tasks;
@@ -19,7 +20,7 @@ namespace FluentlyHttpClient
 	/// Fluent HTTP request, which wraps the <see cref="HttpRequestMessage"/> and add additional features.
 	/// </summary>
 	[DebuggerDisplay("{DebuggerDisplay,nq}")]
-	public class FluentHttpRequest
+	public class FluentHttpRequest : IFluentHttpMessageState
 	{
 		private string DebuggerDisplay => $"[{Method}] '{Uri}'";
 
@@ -62,16 +63,22 @@ namespace FluentlyHttpClient
 		public CancellationToken CancellationToken { get; set; }
 
 		/// <summary>
-		/// Gets or sets a key/value collection that can be used to share data within the scope of request/response.
+		/// Gets or sets a key/value collection that can be used to share data within the scope of request/response or middleware.
 		/// </summary>
-		public IDictionary<object, object> Items { get; protected set; } = new Dictionary<object, object>();
+		public IDictionary<object, object> Items { get; protected set; }
+
+		/// <summary>
+		/// Formatters to be used for content negotiation for "Accept" and also sending formats. e.g. (JSON, XML)
+		/// </summary>
+		public MediaTypeFormatterCollection Formatters { get; set; }
 
 		/// <summary>
 		/// Initializes a new instance.
 		/// </summary>
-		public FluentHttpRequest(HttpRequestMessage message)
+		public FluentHttpRequest(HttpRequestMessage message, IDictionary<object, object> items = null)
 		{
 			Message = message;
+			Items = items ?? new Dictionary<object, object>();
 		}
 
 		/// <summary>